+++ packages/generator/src/utils/get-blitz-dependency-version.ts
@@ -1,7 +1,7 @@
 import {logFailedVersionFetch} from "./get-latest-version"
 import {fetchDistTags} from "./npm-fetch"
 
-const CURRENT_BLITZ_TAG = "alpha"
+const CURRENT_BLITZ_TAG = "latest"
 
 export const getBlitzDependencyVersion = async () => {
   try {