+++ packages/blitz-next/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @blitzjs/next
 
+## 2.0.0-beta.2
+
+### Patch Changes
+
+- db7233db: Bump react, react-dom, @types/react and next versions
+
+  This fixes a console warning: `Warning: Received`true`for a non-boolean attribute`global`.` when using `styled-jsx`. Versions bump also fixes React Hydration error that happens on and off when using `redirectAuthenticatedTo`.
+
+- Updated dependencies [db7233db]
+  - @blitzjs/rpc@2.0.0-beta.2
+
+## 2.0.0-beta.1
+
+### Patch Changes
+
+- 0f4926fd: Set current Blitz tag to latest
+- Updated dependencies [0f4926fd]
+  - @blitzjs/rpc@2.0.0-beta.1
+
+## 2.0.0-beta.72
+
+### Patch Changes
+
+- 2cc888ef: Beta release
+- Updated dependencies [2cc888ef]
+  - @blitzjs/rpc@2.0.0-beta.72
+
+## 2.0.0-alpha.71
+
+### Patch Changes
+
+- @blitzjs/rpc@2.0.0-alpha.71
+
 ## 2.0.0-alpha.70
 
 ### Patch Changes